Se arregló almacenamiento de los tipos de moneda por error de dedo en creación de tabla en base de datos. Ahora se muestra el tipo de moneda de la cuenta en la lista de cuentas. Se arregló que solo se mostraba un elemento de la lista de cuentas. Ahora para el constructor de cada POJO, se llenan los datos con las instrucciones internas del POJO.

+++ app/src/main/java/mx/shellcore/android/micontador/utils/UtilsCurrency.java
@@ -6,7 +6,7 @@ public class UtilsCurrency {
         return (long) num;
     }
 
-    public static double getFractionalPart(double num) {
-        return ((num - getIntegerPart(num)) * 100);
+    public static long getFractionalPart(double num) {
+        return (long) ((num - getIntegerPart(num)) * 100);
     }
 }
